[Ignore] Added a weapon option for torpedos to make them more effective
Cruise missiles now have the option to differentiate between surface and UW targets. If a torpedo is targeting a UW with it on, it will remain cruisealt elmos above the sea floor. If a torpedo is targeting a surface target, it will want to be cruisealt elmos below the surface.