[Ignore] [Ban] mr0neoncat
mr0neoncat has been banned for concern trolling (we do not wish to condone this behavior in our community as we must focus our efforts into improving the game) as well as attempting to defame the lead developer by claiming that I "do not like people playing base game". This behavior is unacceptable, and will not be tolerated. This ban will remain until a sincere apology is written.